+++ clang/test/Sema/warn-fortify-source.c
@@ -75,6 +75,12 @@ void call_strlcat(void) {
   __builtin_strlcat(s1, s2, 20); // expected-warning {{'strlcat' size argument is too large; destination buffer has size 10, but size argument is 20}}
 }
 
+void call_strlcpy(void) {
+  char s1[10], s2[20];
+  __builtin_strlcpy(s2, s1, 20);
+  __builtin_strlcpy(s1, s2, 20); // expected-warning {{'strlcpy' size argument is too large; destination buffer has size 10, but size argument is 20}}
+}
+
 void call_strcpy(void) {
   const char *const src = "abcd";
   char dst[4];